order-service/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
mysqlclient==2.2.4
requests==2.31.0
django-cors-headers==4.3.1
